requirements.txt
absl-py==2.3.0
alembic==1.16.2
aniso8601==10.0.1
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
Authlib==1.6.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
curl_cffi==0.11.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
email-validator==1.1.3
eventkit==1.0.3
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
finta==1.3
Flask==2.0.2
Flask-APScheduler==1.13.1
flask-cors==6.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
Flask-OAuthlib==0.9.6
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==25.2.10
fonttools==4.58.0
frozendict==2.4.6
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.72.1
gunicorn==20.1.0
h5py==3.14.0
hmmlearn==0.3.3
htmlmin==0.1.12
htmlminf==0.1.13
ib-insync==0.9.86
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.0.2
joblib==1.5.1
jsmin==3.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.10.0
kiwisolver==1.4.8
lesscpy==0.15.1
libclang==18.1.1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multitasking==0.0.11
namex==0.1.0
narwhals==1.41.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
oauthlib==2.1.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
pandas-ta==0.3.14b0
parameterized==0.9.0
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.1
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rcssmin==1.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.1.0
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
secure-smtplib==0.1.1
shap==0.47.2
six==1.17.0
slicer==0.0.8
soupsieve==2.7
SQLAlchemy==1.4.29
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.0.3
wrapt==1.17.2
WTForms==3.0.0
xxhash==3.5.0
yfinance==0.2.62
django-ratelimit